requirements.txt
asgiref==3.3.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
Django==3.1.3
django-cleanup==5.1.0
django-cors-headers==3.5.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
docx==0.2.4
idna==2.10
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==2.11.2
lxml==4.6.2
MarkupSafe==1.1.1
oauthlib==3.1.0
Pillow==8.0.1
pycparser==2.20
PyJWT==1.7.1
python-docx==0.8.10
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
smtp==0.1.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
zipp==3.4.0
